complete aws, python, javascript test
